+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Update the Markdown syntax to improve the highlighting of code blocks.
 - Update the Go syntax to fix some highlighting keywords (thanks to Karam!).
 - Improve the stability.
+- [trivial] Improve the tooltip appearance when hovering over a truncated item in the file browser.
 
 
 ### Fixes